requirements.txt
numpy>=1.24.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
aiohttp>=3.8.0
tqdm>=4.65.0
sentence-transformers>=2.2.2
